core_api/requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.0
fastapi-cli==0.0.7
greenlet==3.1.1
h11==0.14.0
hiredis==3.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.4
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.16
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.1
pydantic-settings==2.3.4
pydantic_core==2.33.0
Pygments==2.19.1
PyMySQL==1.1.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.0.6
rich==14.0.0
rich-toolkit==0.14.1
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.0
ujson==5.10.0
uvicorn==0.30.1
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
